+++ src/deli/design/hamming.py
